Sync proxying of GitHub Pages with the website proxy
This, for the moment, has the GitHub Pages build work both as a
pages site _and_ under proxy (hence stripping /website/ from the
pages in the proxy). If we're happy to let the GitHub Pages site
not work unproxied, we can drop that.

The motivation for keeping both working is a known issue where some
forms of url will get redirected such that they send the user to the
underlying GitHub Pages site. While we should definitely fix that
(and the inclusion of an extra `proxy_redirect` line in the new
location block tries to do exactly that), keeping that experience
seem to work for users feels worth it.